Moved PersonsSearch to components folder
