+++ src/pages/Punch/ClearPunch/useClearPunchFacade.tsx
@@ -67,7 +67,6 @@ const useClearPunchFacade = (
     const [clearPunchStatus, setClearPunchStatus] = useState(
         AsyncStatus.INACTIVE
     );
-    const source = Axios.CancelToken.source();
 
     const updateDatabase = async (
         endpoint: UpdatePunchEndpoint,
@@ -237,6 +236,7 @@ const useClearPunchFacade = (
     };
 
     useEffect(() => {
+        const source = Axios.CancelToken.source();
         (async (): Promise<void> => {
             try {
                 const [
@@ -292,7 +292,6 @@ const useClearPunchFacade = (
         handleEstimateChange,
         showPersonsSearch,
         setShowPersonsSearch,
-        cancelToken: source.cancel,
     };
 };
